+++ src/main/java/com/yuuko/core/commands/media/commands/GithubCommand.java
@@ -28,7 +28,7 @@ public class GithubCommand extends Command {
     public void onCommand(MessageEvent e) {
         String[] commandParameters = e.getParameters().split("\\s+", 2);
 
-        final String url = BASE_URL + Sanitiser.scrubString(commandParameters[0], true) + "/" + Sanitiser.scrubString(commandParameters[1], true) + "?access_token=" + Utilities.getApiKey("github");
+        final String url = BASE_URL + Sanitiser.scrub(commandParameters[0], true) + "/" + Sanitiser.scrub(commandParameters[1], true) + "?access_token=" + Utilities.getApiKey("github");
         final JsonObject json = new RequestHandler(url).getJsonObject();
 
         if(json == null) {