+++ src/main/java/com/yuuko/core/commands/media/commands/KitsuCommand.java
@@ -23,7 +23,7 @@ public class KitsuCommand extends Command {
     @Override
     public void onCommand(MessageEvent e) {
         try {
-            final String url = BASE_URL + Sanitiser.scrubString(e.getParameters(), true) + "&page[limit]=1";
+            final String url = BASE_URL + Sanitiser.scrub(e.getParameters(), true) + "&page[limit]=1";
             JsonObject json = new RequestHandler(url, new RequestProperty("Accept", "application/vnd.api+json"), new RequestProperty("Content-Type","application/vnd.api+json")).getJsonObject();
 
             if(json == null || json.getAsJsonArray("data").size() < 1) {