+++ src/main/java/com/yuuko/core/io/RequestHandler.java
@@ -7,6 +7,8 @@ import com.yuuko.core.io.entity.RequestProperty;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -69,7 +71,7 @@ public class RequestHandler {
     /**
      * Retrieves the content as a JsonObject which can be handled and manipulated with the Google Gson package.
      *
-     * @return JsonObject
+     * @return {@link JsonObject}
      * @throws IllegalStateException IllegalStateException
      */
     public JsonObject getJsonObject() throws IllegalStateException {
@@ -84,12 +86,26 @@ public class RequestHandler {
     /**
      * Retrieves the content as a JsonArray which can be handled and manipulated with the Google Gson package.
      *
-     * @return JsonArray
+     * @return {@link JsonArray}
      * @throws IllegalStateException IllegalStateException
      */
     public JsonArray getJsonArray() throws IllegalStateException {
         try {
             return (content == null) ? null : new JsonParser().parse(content).getAsJsonArray();
+        } catch(Exception ex) {
+            log.error("An error occurred while running the {} class, message: {}", this, ex.getMessage(), ex);
+            return null;
+        }
+    }
+
+    /**
+     * Retrieves the content as a Jsoup Document.
+     *
+     * @return {@link Document}
+     */
+    public Document getDocument() {
+        try {
+            return (content == null) ? null : Jsoup.parse(content);
         } catch(Exception ex) {
             log.error("An error occurred while running the {} class, message: {}", this, ex.getMessage(), ex);
             return null;